+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## [1.0.1] - 2022-12-03
+## [1.0.2] - 2022-01-02
+
+### Added
+- StringBuilder conditional append
+
+### Changed
+- Updated namespaces in most files
+
+## [1.0.1] - 2022-01-02
 
 ### Added
 - Int extensions